Create bridge between Arduino and Console
Found an example script in the documentation for the pyserial package.
It works fine without any changes. Doesn't have to waste time converting
anything. The byte stream is just ingested by the program and sent out
to the console via a tcp socket. Next -- make an html page for a config
GUI so user can easily set the network addres and serial port.