Populate form fields with current settings on load
Added methods to the form classes to put the values of the current
settings in their respective form fields when the settings page is
loaded. Need to program how to receive new settings when the form
is submitted. Need to cut up the serial bridge script to feed it info
from config.ini