Run two processes without blocking each other
Start the thread for the osc bridge before calling the flask app to
serve the settings page. The flask app is threaded by default, so didn't
need to be threaded in the main script. As long as the bridge thread is
started before the flask app is ran, the program works correctly.

There is a ~5 second delay before the osc device is connected to the
console. I thought that with the original sample script that the
connection was instantaneous, but now that script also has a delay. Not
the worst thing, but also not sure what the cause is.

TODO: When new settings are applied, the script needs to restart to read
the new config data.