Define serial and socket connections in main.py
Define the connections in the main program instead of the oscbridge
file. It should now be possible to close the connections and rerun
main().

TODO: Create a way to shut down the bridge and the flask app before
calling main() again. This should be able to be called in the config
app when the form is submitted in order to apply the new user settings