Create run() and restart() functions to handle restarting
Program now succesffuly restarts when the form is submitted on the
settings page, but breaks down when settings are changed or the
'apply changes' button is pressed multiple times.

TODO: Flesh out error checking in the open_bridge function to find out
where the flow is getting hung up. It seems like if the socket or
serial connection fail, the process isn't fully shutdown.