Port program to Raspberry Pi Zero 2
Moved program over to the Raspberry Pi Zero 2. Changed the path of
"config.ini" to a full path because the rpi couldn't see it as
a relative path. Set up a systemd service to start main.py.
 Everything seems to be working properly.

TODO: Add a shutdown script. A custom osc message can be sent
from the console to trigger the shutdown. The bridge should be
properly shut down before calling the "shutdown" system command.